Added Version Options
Added the ability to toggle between original and new versions. The old
script is contained in a new script file called script_original.rpy.
Changes can be made to the rest of the script files without affecting
the original. Any changes to the assets or pointers will change the
original, though. If we want to change anything asset-wise, we should
make new versions instead of changing the old ones.